refactor: move flow elements out of pages/index
